Chapter 11 - Fix bug in types
Theorem `no` is incorrect and therefore does not (presently)
typecheck.

Here's the error:

```
- + Errors (1)
 `-- Ex_9_1_6.idr line 35 col 30:
     no notFound LastOne is a valid case
```

We can see that the type of `no` is wrong. For replace `no` with
the folowing code, for when the implementation of `no` is removed
(to make it typecheck), we can define the following value:

```
doesNotMakeSense : Void
doesNotMakeSense = no nilHasNoLast (LastOne {value = 42})
```

This demonstrates that `no`, were it to typecheck, would make our
system inconsistent.

The solution is to make `no` apply only to non-empty lists.